Add files via upload
This study aims to contribute to the improvement of academic dropout by exploring the 
application of machine learning techniques to train a classifier that will predict if a student is 
likely to drop out. The model can be used by academic institutions to identify and support 
students at risk and provide timely support intervention.